hello-py/pyproject.toml
[tool.poetry]
name = "hello"
version = "0.1.0"
description = ""
authors = ["aita <792803+aita@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7"
bottle = "^0.12.17"
grpcio = "^1.23"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.23"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
greeter = 'hello.greeter_server:serve'
gateway = 'hello.gateway:serve'
